03-multimodal/pyproject.toml
[project]
name = "ai-assistant-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.7",
    "openai>=1.30",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
